+++ src/functions/quota-collector/main.py
@@ -16,7 +16,10 @@ from modules.qmchecks.chime import get_current_quotastatus_chime
 from modules.qmchecks.interconnect import get_current_quotastatus_interconnect
 from modules.qmchecks.rtbfabric import get_current_quotastatus_rtbfabric
 from modules.qmchecks.tnb import get_current_quotastatus_tnb
+from modules.qmchecks.dax import get_current_quotastatus_dax
 from modules.qmchecks.drs import get_current_quotastatus_drs
+from modules.qmchecks.license_manager_user_subscriptions import (
+    get_current_quotastatus_license_manager_user_subscriptions)
 from modules.qmchecks.schemas import get_current_quotastatus_schemas
 from modules.qmchecks.launchwizard import get_current_quotastatus_launchwizard
 from modules.qmchecks.vpc.vpc import get_current_quotastatus_vpc
@@ -238,7 +241,10 @@ def lambda_handler(event, context):
     entries.extend(get_current_quotastatus_interconnect(ctx=ctx, skip=official))
     entries.extend(get_current_quotastatus_rtbfabric(ctx=ctx, skip=official))
     entries.extend(get_current_quotastatus_tnb(ctx=ctx, skip=official))
+    entries.extend(get_current_quotastatus_dax(ctx=ctx, skip=official))
     entries.extend(get_current_quotastatus_drs(ctx=ctx, skip=official))
+    entries.extend(get_current_quotastatus_license_manager_user_subscriptions(
+        ctx=ctx, skip=official))
     entries.extend(get_current_quotastatus_schemas(ctx=ctx, skip=official))
     entries.extend(get_current_quotastatus_launchwizard(ctx=ctx, skip=official))
     entries.extend(get_current_quotastatus_account_services(ctx, skip=official))